Fix: impose some corrections following warnings from gym env

+++ src/civrealm/envs/freeciv_wrapper/utils.py
@@ -31,7 +31,7 @@ def resize_data(data: np.ndarray, size: int):
     remain_shape = data.shape[1:]
     data = data.copy()
     data.resize([size, *remain_shape], refcheck=False)
-    return data
+    return data.astype(np.int32)
 
 
 def add_shape(shape_1, shape_2):
